requirements.txt
flask
openai-whisper
tensorflow
librosa
numpy
pydub
